wire-subsystem: Extract the interpreter out of NotificationSubsystem
Imports are nicer this way
